requirements.django.txt
Django>=1.11
djangorestframework==3.4.0
coreapi
